backend/requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0

# CORS
django-cors-headers==4.3.1

# Environment variables
python-decouple==3.8

# Google OAuth
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# JWT
PyJWT==2.8.0

# Utils
pytz==2024.1